Move execute_tx to SuiClient
